### 2022-09-28 #4802 @blebon
There are only two OpenFOAM reviewers on the spreadsheet:

- li.dy@dyfluid.com (I doubt their username is really `Yes` as mentionned in the spreadsheet. It seems to be https://github.com/dyfluid instead).
- [ctdegroot](https://github.com/ctdegroot )

This is an active OpenFOAM developer but he is not on the reviewer list:
- [ancolli](https://github.com/ancolli) Disclosure: he did contact me at one point regarding my acoustic streaming application but I had limited interaction with him. Only a few emails.

These are active researchers working on solidification with OpenFOAM and on Github. They are not on the reviewers list:
- [atharvajoshi05](https://github.com/atharvajoshi05)
- [viktorlindstroemEMPA](https://github.com/viktorlindstroemEMPA)

Otherwise, these are more active CFD reviewers from the list, but they don't seem to work on OpenFOAM:
- [chennachaos](https://github.com/chennachaos)
- [jayten](https://github.com/jayten)
- [bzindovic](https://github.com/bzindovic)
- [mikaem](https://github.com/mikaem)
- [jramirezcr](https://github.com/jramirezcr)
