+++ paper/paper.md
@@ -20,31 +20,31 @@ bibliography: paper.bib
 
 # Summary
 
-Direct-chill (DC) casting is a semi-continuous casting process that is used for producing aluminium and magnesium alloy billets [Eskin2008]. As illustrated in (\autoref{fig:DC_Sketch}), the process consists of feeding melt (liquid metal) into a mould containing a movable bottom (the ram). The ram is lowered and the billet is pulled downwards by gravity. As the billet exists the mould, its exposed outer surface is chilled with water jets to hasten cooling (hence the term direct-chill).
+Direct-chill (DC) casting is a semi-continuous casting process that is used for producing aluminium and magnesium alloy billets [@Eskin2008]. As illustrated in (\autoref{fig:DC_Sketch}), the process consists of feeding melt (liquid metal) into a mould containing a movable bottom (the ram). The ram is lowered and the billet is pulled downwards by gravity. As the billet exists the mould, its exposed outer surface is chilled with water jets to hasten cooling (hence the term direct-chill).
 
 ![The direct-chill casting process.\label{fig:DC_Sketch}](../docs/source/images/DC_Schematic.png){ width=40% }
 
 After DC casting, the billets are subject to further energy intensive processing (e.g. extrusion, rolling, forging ...). If billets contain defects or are of poor quality, they are either rejected or will ultimately result in components of poor quality: this makes DC casting a crucial step in alloy processing. In a push for lower carbon emissions and greater recyclability of metallic components, DC casting is currently the subject of various optimization studies involving the presence of external fields. These studies include:  
 
-1. the use of ultrasound [Eskin2014; Lebon2019a; Lebon2019b; Subroto2021; Yamamoto2021].
-2. high-shear melt conditioning with a rotor-stator mixer [Xia2013; PrasadaRao2014; Li2017; Lebon2018; Lebon2020; Sree2021], and 
-3. electromagnetic fields [Hatic2018].
+1. the use of ultrasound [@Eskin2014; @Lebon2019a; @Lebon2019b; @Subroto2021; @Yamamoto2021].
+2. high-shear melt conditioning with a rotor-stator mixer [@Xia2013; @PrasadaRao2014; @Li2017; @Lebon2018; @Lebon2020; @Sree2021], and 
+3. electromagnetic fields [@Hatic2018].
 
 All these studies benefit from numerical modeling to cut down on the number of experiments that are required to study the process.
  
 # Statement of need
 
-`directChillFoam` is an OpenFOAM solver [OpenFOAM] for the computational fluid dynamics (CFD) modelling of the DC casting process. `directChillFoam` extended OpenFOAM's `buoyantPimpleFoam` solver to include the following functionalities:  
+`directChillFoam` is an OpenFOAM solver [@OpenFOAM] for the computational fluid dynamics (CFD) modelling of the DC casting process. `directChillFoam` extended OpenFOAM's `buoyantPimpleFoam` solver to include the following functionalities:  
 
-* An improved solidification model that can handle alloys with more than two components and with a liquid fraction-temperature profile that can be entered with an interpolation table. These enable the use of thermophysical properties that can be calculated on the fly by a CALPHAD package e.g. [Andersson2002].
+* An improved solidification model that can handle alloys with more than two components and with a liquid fraction-temperature profile that can be entered with an interpolation table. These enable the use of thermophysical properties that can be calculated on the fly by a CALPHAD package e.g. [@Andersson2002].
 * Switching between mushy and slurry zones in the phase transition region to accurately handle the flow near the solidification front.
 * Treatment of the secondary cooling heat transfer boundary condition.
-* An energy corrector loop to correctly update the phase fraction while conserving energy, based on the approach of [Faden2018].
+* An energy corrector loop to correctly update the phase fraction while conserving energy, based on the approach of [@Faden2018].
[…]
