### 2022-09-28 #4802 @blebon
Fixed references and resized DC casting figure.
