+++ .github/ISSUE_TEMPLATE.md
@@ -1,3 +1,0 @@
-👋👋 Do you want to contribute to blebon/directChillFoam? 👋👋
-
-If you have discovered a bug or would like to make a request, please browse the open issues before opening a new one. You might also like to have a look at the [Open Source Guide](https://opensource.guide/how-to-contribute/).
