+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,31 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: "[BUG]"
+labels: ''
+assignees: ''
+
+---
+
+**First verify  that the bug has not already been reported.**
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+Cosmetic changes (spaces/tabs ...) will be rejected.
+
+**To Reproduce**
+Steps to reproduce the behavior:  
+Submit an **example case** demonstrating the unexpected behaviour.
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots, contour/vector plots, graphs, or any other visual aid to help explain your problem.
+
+**Desktop (please complete the following information):**
+ - OS: [e.g. Ubuntu]
+ - Version [e.g. OF9]
+
+**Additional context**
+Add any other context about the problem here.
