+++ .github/PULL_REQUEST_TEMPLATE/pull_request_template.md
@@ -1,0 +1,13 @@
+
+👋👋 Thank you for wanting to contribute to directChillFoam. 👋👋
+
+* [ ] Before submitting, please read the [How to Contribute to OpenFOAM guide](https://openfoam.org/dev/how-to-contribute/). Adhere to OpenFOAM's coding convention.
+* [ ] Does your contribution pass the tutorial validation (predicted temperature matches experimental data)?
+
+#### **Did you write a bug fix patch?**
+
+* [ ] Describe the problem and the solution as clearly as possible.
+
+#### **Is you fix purely cosmetic?**
+
+* [ ] Cosmetic changes (spaces/tabs ...) will be rejected.
