+++ .github/workflows/sphinx_docs_to_gh_pages.yaml
@@ -17,7 +17,7 @@ jobs:
         with:
           fetch-depth: 0
       - name: Make conda environment
-        uses: conda-incubator/setup-miniconda@v3
+        uses: conda-incubator/setup-miniconda@v2
         with:
           python-version: 3.9    # Python version to build the html sphinx documentation
           environment-file: .github/conda-envs/docs_env.yaml    # Path to the documentation conda environment
