+++ .github/workflows/sphinx_docs_to_gh_pages.yaml
@@ -24,10 +24,6 @@ jobs:
           auto-update-conda: false
           auto-activate-base: false
           show-channel-urls: true
-      - name: Installing the library
-        shell: bash -l {0}
-        run: |
-          python setup.py install
       - name: Running the Sphinx to gh-pages Action
         uses: uibcdf/action-sphinx-docs-to-gh-pages@v1.0.0
         with:
