+++ .github/workflows/sphinx_docs_to_gh_pages.yaml
@@ -28,7 +28,7 @@ jobs:
         uses: uibcdf/action-sphinx-docs-to-gh-pages@v1.0.0
         with:
           branch: master
-          dir_docs: docs
+          dir_docs: docs/source
           sphinxapiopts: '--separate -o . ../'
           sphinxapiexclude: '../*setup* ../*.ipynb'
           sphinxopts: ''
