### 2022-09-29 8b0eed2
Github actions wmake
