+++ .github/workflows/wmake.yml
@@ -24,7 +24,7 @@ jobs:
     runs-on: ubuntu-latest
     name: wmake
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/cache@v3
 
     - name: Install OpenFOAM 9
       shell: bash
@@ -47,7 +47,7 @@ jobs:
       working-directory: src/fvConstraints
       run: |
         source /opt/openfoam9/etc/bashrc
-        /opt/openfoam9/wmake/wmake libso
+        wmake libso
     
     - name: wmake fvModels
       shell: bash
