### 2022-09-29 f008b6a
Github actions wmake
