+++ docs/source/conf.py
@@ -13,12 +13,14 @@
 # import os
 # import sys
 # sys.path.insert(0, os.path.abspath('.'))
+from datetime import datetime
 
 
 # -- Project information -----------------------------------------------------
 
 project = 'directChillFoam'
-copyright = '2022, Bruno Lebon'
+year = datetime.now().year
+copyright = f'{year}, Bruno Lebon'
 author = 'Bruno Lebon'
 
 # The full version, including alpha/beta/rc tags
@@ -55,3 +57,10 @@ html_theme = 'alabaster'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+
+html_theme_options = {
+    'github_user': 'blebon',
+    'github_repo': 'directChillFoam',
+    'github_banner': True,
+    'sidebar_width': '300px',
+}
