+++ .github/workflows/sphinx_docs_to_gh_pages.yaml
@@ -5,7 +5,7 @@ on:
     branches:
       - master
     paths:
-      - 'paper/**'
+      - 'docs/**'
   workflow_dispatch:
 
 jobs:
