+++ .github/conda-envs/docs_env.yaml
@@ -16,3 +16,4 @@ dependencies:
   - nbsphinx
   - recommonmark
   - sphinx-markdown-tables
+  - breathe

[…]
