+++ .github/workflows/sphinx_docs_to_gh_pages.yaml
@@ -6,20 +6,25 @@ on:
       - master
     paths:
       - 'docs/**'
+      - 'doxygen/**'
   workflow_dispatch:
 
 jobs:
   sphinx_docs_to_gh-pages:
     runs-on: ubuntu-latest
     name: Sphinx docs to gh-pages
+
     steps:
       - uses: actions/checkout@v2
         with:
           fetch-depth: 0
+
       - name: Check doxygen is installed and latest version
         run: |
           sudo apt-get update
+          sudo apt-get -y install graphviz libclang1-9 libclang-cpp9
           sudo apt-get -y install doxygen
+          
       - name: Make conda environment
         uses: conda-incubator/setup-miniconda@v2
         with:
@@ -28,6 +33,13 @@ jobs:
           auto-update-conda: false
           auto-activate-base: false
           show-channel-urls: true
+      
+      - name: Build doxygen XML
+        working-directory: doxygen
+        run: |
+          make clean
+          make all
+
       - name: Running the Sphinx to gh-pages Action
         uses: uibcdf/action-sphinx-docs-to-gh-pages@v1.0.0
         with:
