+++ docs/source/conf.py
@@ -18,9 +18,9 @@ from datetime import datetime
 # Breathe
 import subprocess
 subprocess.call("make clean", shell=True)
-subprocess.call(f"cd ..{sep}..{sep}doxygen ; doxygen Doxyfile", shell=True)
+subprocess.call(f"doxygen Doxyfile", shell=True)
 
-breathe_projects = { "directChillFoam": "../build/xml" }
+breathe_projects = { "directChillFoam": f"..{sep}build{sep}xml" }
 breathe_default_project = "directChillFoam"
 
 
