+++ doxygen/Makefile
@@ -1,10 +1,0 @@
-# Minimal makefile for Doxygen documentation
-#
-
-.PHONY: all clean
-
-all:
-	doxygen Doxyfile
-
-clean:
-	rm -rf html ../docs/source/xml
