+++ docs/source/conf.py
@@ -17,8 +17,7 @@ from datetime import datetime
 
 # Breathe
 import subprocess
-subprocess.call("make clean", shell=True)
-subprocess.call(f"doxygen Doxyfile", shell=True)
+subprocess.call(f"cd ..; doxygen Doxyfile", shell=True)
 
 breathe_projects = { "directChillFoam": f"..{sep}build{sep}xml" }
 breathe_default_project = "directChillFoam"
