+++ .github/workflows/sphinx_docs_to_gh_pages.yaml
@@ -32,13 +32,6 @@ jobs:
           auto-update-conda: false
           auto-activate-base: false
           show-channel-urls: true
-      
-      - name: Build doxygen XML
-        working-directory: docs
-        continue-on-error: true
-        run: |
-          make clean
-          doxygen Doxyfile
 
       - name: Running the Sphinx to gh-pages Action
         uses: uibcdf/action-sphinx-docs-to-gh-pages@v1.0.0
