### 2022-10-04 04f9f87
Removed obsolete doxygen tags
