+++ .github/workflows/sphinx_docs_to_gh_pages.yaml
@@ -45,6 +45,7 @@ jobs:
 
       - name: Running the Sphinx to gh-pages Action
         uses: uibcdf/action-sphinx-docs-to-gh-pages@v1.0.0
+        shell: bash
         with:
           branch: master
           dir_docs: docs/source
