### 2022-10-04 73d407a
Copy doxyFilter from OF source
