### 2022-10-04 bdab61f
Fix references doxygen formatting
