### 2022-10-04 f9c2f04
Fix references doxygen formatting
