### 2022-10-05 1a0a4b3
Python script documentation
