+++ tutorials/heatTransfer/directChillFoam/Vreeman2002/write_htc.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-"""Writes HTC table file in foam format
+"""Writes HTC table file in foam format.
 
 This script computes the HTC for the water film, including the
 effect of nucleate boiling. Outputs the HTC table in foam
@@ -26,21 +26,37 @@ from numpy import array, isnan, abs, linspace, e, power, pi, sqrt
 from scipy.interpolate import UnivariateSpline, interp1d
 
 __author__ = "Bruno Lebon"
-__copyright__ = "Copyright, 2021, Brunel University London"
+__copyright__ = "Copyright, 2022, Brunel University London"
 __credits__ = ["Bruno Lebon"]
 __email__ = "Bruno.Lebon@brunel.ac.uk"
 __status__ = "Production"
 
 
 def rohsenhow(T=930, n=1.0, g0=1.0, q_water=90.0, D=0.255, T_water=372.8, T_amb=328.5):
-    """HTC from water film"""
+    """Rohsenhow correlation.
+
+    :param T: Billet temperature [K].
+    :type T: float
+    :param n: Experimental constant equal to 1 for water and 1.7 for other fluids.
+    :type n: float
+    :param g0: Force conversion factor [kg m/N s^2].
+    :type g0: float
+    :param q_water: Water flow rate [L/min].
+    :type q_water: float
+    :param D: Diameter [m].
+    :type D: float
+    :param T_water: Temperature of water jet [K].
+    :type T_water: float
+    :param T_amb: Temperature of the lab [K].
+    :type T_amb: float
+    :return: 0 if successful.
+    :rtype: int
+    """
     mu = 2.829e-4
     k = 0.6790
     rho = 958.6
     g = 9.81
     Cp = 4215.0
-    # flow_rate = 0.00133333 # Volume flow rate
-    # flow_rate = 90 * 0.001 / 60
     flow_rate = q_water * 0.001 / 60.0
     D = 0.255
     Gamma = flow_rate / (pi * D)
@@ -48,7 +64,7 @@ def rohsenhow(T=930, n=1.0, g0=1.0, q_water=90.0, D=0.255, T_water=372.8, T_amb=
     sigma = 5.9e-2
     Pr = Cp * mu / k
     C_f = 0.011
-    A = power(mu ** 2 / (k ** 3 * rho ** 2 * g), -1 / 3)
+    A = power(mu**2 / (k**3 * rho**2 * g), -1 / 3)
     B = power(Pr, 1 / 3)
     C = power(4 * Gamma / mu, 1 / 3)
     h_conv = 0.01 * B * C * A
@@ -61,14 +77,26 @@ def rohsenhow(T=930, n=1.0, g0=1.0, q_water=90.0, D=0.255, T_water=372.8, T_amb=
                 mu
                 * Hfg
                 * sqrt(g * (rho - 0.5903) / (g0 * sigma))
-                * power(Cp * (T - T_water) / (C_f * Hfg * Pr ** n), 3)
+                * power(Cp * (T - T_water) / (C_f * Hfg * Pr**n), 3)
             )
             h = (q_A + q_conv) / (T - T_amb)
     return h
 
 
 def weckmann_niessen(T, q_water=90.0, D=0.255, T_water=372.8):
-    """Weckmann-Niessen correlation"""
+    """Weckmann-Niessen correlation.
+
+    :param T: Billet temperature [K].
+    :type T: float
+    :param q_water: Water flow rate [L/min].
+    :type q_water: float
+    :param D: Diameter [m].
+    :type D: float
+    :param T_water: Temperature of water jet [K].
+    :type T_water: float
+    :return: 0 if successful.
+    :rtype: int
+    """
     A = -167000.0
     B = 352
[…]
