+++ docs/source/tutorials/modules.rst
@@ -1,0 +1,33 @@
+============================
+Python scripts documentation
+============================
+
+VTK conversion
+==============
+
+.. automodule:: Vreeman2002.convert_vtk
+   :members:
+
+PVD generation
+==============
+
+.. automodule:: Vreeman2002.generate_pvd
+   :members:
+
+HTC file creation
+=================
+
+.. automodule:: Vreeman2002.write_htc
+   :members:
+
+Temperature validation
+======================
+
+.. automodule:: Vreeman2002.plot_line
+   :members:
+
+Sump contour
+============
+
+.. automodule:: Vreeman2002.plot_sump
+   :members:

[…]
