### 2022-10-05 3bb8c90
Missing dependencies for tutorial scripts
