### 2022-10-05 d100bbc
Missing dependencies for tutorial scripts
