+++ .github/conda-envs/docs_env.yaml
@@ -20,3 +20,7 @@ dependencies:
   - numpy
   - natsort
   - matplotlib
+  - vtk
+  - pandas
+  - seaborn
+  - scipy
