### 2022-10-05 a2e4784
Fixed doi link to continuum mixture model reference
