### 2022-10-21 #4802 @meg-simula
:wave:  @ancolli, @chennachaos, @jayten, @bzindovic, @florianwechsung would you be available to review this submission for JOSS? Your expertise and contribution would be much appreciated. Do take a look at the JOSS reviewer guidelines here:

https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

and please respond in this pre-review issue.
