### 2022-10-21 #4802 @ancolli
Sorry, I do not have time!

El vie, 21 oct 2022 a las 13:41, Marie E Rognes ***@***.***>)
escribió:

> 👋 @ancolli <https://github.com/ancolli>, @chennachaos
> <https://github.com/chennachaos>, @jayten <https://github.com/jayten>,
> @bzindovic <https://github.com/bzindovic>, @florianwechsung
> <https://github.com/florianwechsung> would you be available to review
> this submission for JOSS? Your expertise and contribution would be much
> appreciated. Do take a look at the JOSS reviewer guidelines here:
>
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> and please respond in this pre-review issue.
>
