### 2022-10-21 #4871 @editorialbot
[REVIEW]: directChillFoam: an OpenFOAM application for direct-chill casting
**Submitting author:** @blebon (<a href="http://orcid.org/0000-0002-9389-1329">Bruno Lebon</a>)
**Repository:** https://github.com/blebon/directChillFoam
**Branch with paper.md** (empty if default branch):
**Version:** OF9.0.1
