+++ tests/test_libs.py
@@ -3,14 +3,22 @@
 
 from ctypes import cdll
 
+
 def test_fvModels():
     """Test that `libmyFvModels.so` can be loaded."""
     cdll.LoadLibrary("libmyFvModels.so")
+
 
 def test_fvConstraints():
     """Test that `libmyFvConstraints.so` can be loaded."""
     cdll.LoadLibrary("libmyFvConstraints.so")
 
+
 def test_ThermophysicalTransportModels():
     """Test that `libmythermophysicalTransportModels.so` can be loaded."""
     cdll.LoadLibrary("libmythermophysicalTransportModels.so")
+
+
+def test_MulticomponentAlloy():
+    """Test that `libmulticomponentAlloy.so` can be loaded."""
+    cdll.LoadLibrary("libmulticomponentAlloy.so")
