+++ .github/workflows/wmake.yml
@@ -23,7 +23,14 @@ jobs:
   build:
     runs-on: ubuntu-latest
     name: wmake
+    strategy:
+      matrix:
+        python-version: ["3.10"]
+    
     steps:
+    - uses: szenius/set-timezone@v1.0
+        with:
+          timezoneLinux: "Europe/London"
     - uses: actions/checkout@v3
 
     # - name: Cache OpenFOAM 9
@@ -99,3 +106,17 @@ jobs:
       run: |
         source /opt/openfoam9/etc/bashrc || true
         wmake
+
+    - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+    
+    - name: Install testing dependencies
+        run: pip install pytest pytest-md pytest-emoji
+    
+    - uses: pavelzw/pytest-action@v2
+      with:
+        emoji: false
+        verbose: true
+        job-summary: true
