+++ .github/workflows/wmake.yml
@@ -108,12 +108,12 @@ jobs:
         wmake
 
     - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
     
     - name: Install testing dependencies
-        run: pip install pytest pytest-md pytest-emoji
+      run: pip install pytest pytest-md pytest-emoji
     
     - uses: pavelzw/pytest-action@v2
       with:
