+++ .github/workflows/wmake.yml
@@ -26,6 +26,8 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.10"]
+    
+    steps:
     - uses: actions/checkout@v3
 
     # - name: Cache OpenFOAM 9
