### 2022-12-30 #4871 @bzindovic
I congratulate the author on an interesting package, I think it will be a great contribution to the overall OpenFOAM ecosystem of solvers. After reading the manuscript and going through the source code, examples and docs, my observations are as follows:

1. State of the field is actually embedded inside the statement of need. Maybe it should be in a separate section?
2. References section includes a reference to the OpenFOAM-dev repository but states the authors as Weller & Jasak (2022). As the repository owner is OpenFOAM Foundation, maybe the reference should cite it as an "author/owner"? Or maybe it was meant to include this reference: Weller, H.G., Tabor, G., Jasak, H., Fureby, C. (1998) A tensorial approach to computational continuum mechanics using object-oriented techniques, Comput. Phys., Volume 12, Issue 6, Pages 620-631?

Once again, I compliment the author for crafting a valuable tool on top of OpenFOAM.
