### 2023-01-03 #4871 @blebon
@bzindovic Happy New Year 2023 and thank you for your review and helpful comments.

1. Thanks for the suggestion. I will separate the state of the field from the statement of need.
2. I will update the reference to OpenFOAM with the suggested paper.

@arfon Happy New Year 2023. Should I implement these suggestions immediately, or should I wait for @chennachaos 's comments before making any change?
