+++ paper/paper.bib
@@ -18,6 +18,19 @@
   year = {2002}
 }
 
+@article{Bennon1987,
+  title = {A continuum model for momentum, heat and species transport in binary solid-liquid phase change systems—I. Model formulation},
+  journal = {International Journal of Heat and Mass Transfer},
+  volume = {30},
+  number = {10},
+  pages = {2161-2170},
+  year = {1987},
+  issn = {0017-9310},
+  doi = {https://doi.org/10.1016/0017-9310(87)90094-9},
+  url = {https://www.sciencedirect.com/science/article/pii/0017931087900949},
+  author = {W.D. Bennon and F.P. Incropera}
+}
+
 @book{Eskin2008,
   author = {Eskin, Dmitry G.},
   booktitle = {Physical Metallurgy of Direct Chill Casting of Aluminum Alloys},
@@ -45,6 +58,29 @@
   title = {{An implicit algorithm for melting and settling of phase change material inside macrocapsules}},
   volume = {117},
   year = {2018}
+}
+
+@article{Hatic2018,
+  author = {Hati{\'{c}}, Vanja and Mavri{\v{c}}, Bo{\v{s}}tjan and Ko{\v{s}}nik, Nejc and {\v{S}}arler, Bo{\v{z}}idar},
+  doi = {10.1016/j.apm.2017.09.034},
+  issn = {0307904X},
+  journal = {Applied Mathematical Modelling},
+  title = {{Simulation of direct chill casting under the influence of a low-frequency electromagnetic field}},
+  volume = {54},
+  year = {2018}
+}
+
+@article{Heyvaert2017,
+  author = {Heyvaert, Laurent and Bedel, Marie and Zalo{\v{z}}nik, Miha and Combeau, Herv{\'{e}}},
+  doi = {10.1007/s11661-017-4238-z},
+  issn = {1543-1940},
+  journal = {Metallurgical and Materials Transactions A},
+  number = {10},
+  pages = {4713--4734},
+  title = {{Modeling of the Coupling of Microstructure and Macrosegregation in a Direct Chill Cast Al-Cu Billet}},
+  url = {https://doi.org/10.1007/s11661-017-4238-z},
+  volume = {48},
+  year = {2017}
 }
 
 @article{Li2017,
@@ -123,6 +159,19 @@
   year = {2017}
 }
 
+@article{Ni1991,
+  author = {Ni, J and Beckermann, C},
+  doi = {10.1007/BF02651234},
+  issn = {1543-1916},
+  journal = {Metallurgical Transactions B},
+  number = {3},
+  pages = {349--361},
+  title = {{A volume-averaged two-phase model for transport phenomena during solidification}},
+  url = {https://doi.org/10.1007/BF02651234},
+  volume = {22},
+  year = {1991}
+}
+
 @misc{OpenFOAM,
   author = {H. Weller and H. Jasak},
   title = {OpenFOAM/OpenFOAM-dev: OpenFOAM Foundation development repository},
@@ -168,16 +217,59 @@
   year = {2021}
 }
 
-@article{Hatic2018,
-  author = {Hati{\'{c}}, Vanja and Mavri{\v{c}}, Bo{\v{s}}tjan and Ko{\v{s}}nik, Nejc and {\v{S}}arler, Bo{\v{z}}idar},
-  doi = {10.1016/j.apm.2017.09.034},
-  issn = {0307904X},
-  journal = {Applied Mathematical Modelling},
-  title = {{Simulation of direct chill casting under the influence of a low-frequency electromagnetic field}},
-  volume = {54},
+@article{Tveito2018,
+  author = {Tveito, Knut Omdal and Pakanati, Akash and M'Hamdi, Mohammed and Combeau, Herv{\'{e}} and Zalo{\v{z}}nik, Miha},
[…]
