+++ paper/paper.md
@@ -30,11 +30,17 @@ After DC casting, the billets are subject to further energy intensive processing
 2. high-shear melt conditioning with a rotor-stator mixer [@Xia2013; @PrasadaRao2014; @Li2017; @Lebon2018; @Lebon2020; @Sree2021], and 
 3. electromagnetic fields [@Hatic2018].
 
-All these studies benefit from numerical modeling to cut down on the number of experiments that are required to study the process.
+All these studies benefit from numerical modelling to cut down on the number of experiments that are required to study the process.
+
+# State of the field
+
+Direct-chill casting modelling broadly falls into two approaches. One class of solution--multiphase or multiple domain models--uses independent conversation equations for each phase, e.g. [@Ni1991]. The conservation equations are then coupled using boundary conditions at phase interfaces, thereby requiring these interfaces to be accurately tracked. More recent implementations of these multiple domain models implement grain motion, e.g. [@Heyvaert2017; @Tveito2018]. This approach is computationally expensive for optimization searches.    
+
+The other approach is based on the continuum model [@Bennon1987] and can also include the effect of free-floating dendrites [@Vreeman2000]. The continuum model avoids the requirement of tracking phase interfaces by implicitly integrating the microscopic description of transport behaviour in a continuum formulation. This approach has been validated with temperature measurements in the sump [@Vreeman2002] and is the starting point for `directChillFoam`.
  
 # Statement of need
 
-`directChillFoam` is an OpenFOAM solver [@OpenFOAM] for the computational fluid dynamics (CFD) modelling of the DC casting process. `directChillFoam` extended OpenFOAM's `buoyantPimpleFoam` solver to include the following functionalities:  
+`directChillFoam` is an OpenFOAM solver [@Weller1998] for the computational fluid dynamics (CFD) modelling of the DC casting process. `directChillFoam` extended OpenFOAM's `buoyantPimpleFoam` solver to include the following functionalities:  
 
 * An improved solidification model that can handle alloys with more than two components and with a liquid fraction-temperature profile that can be entered with an interpolation table. These enable the use of thermophysical properties that can be calculated on the fly by a CALPHAD package e.g. [@Andersson2002].
 * Switching between mushy and slurry zones in the phase transition region to accurately handle the flow near the solidification front.
@@ -44,10 +50,10 @@ All these studies benefit from numerical modeling to cut down on the number of e
 
 `directChillFoam` makes the most of OpenFOAM's modularity and implements each significant functionality within reusable libraries. The power of `directChillFoam` lies within the ease with which the solver can be extended. For example, `directChillFoam` has been coupled with a novel acoustic streaming solver [@acousticStreamingFoam] to model acoustic streaming inside a direct-chill casting mould [@Lebon2019b]. Such coupling of a newly published complex numerical method would be non-trivial in a commercial CFD software package.
 
-`directChillFoam` was designed to be used by both academics involved in fundamental casting research and industry practitioners for optimization problems. It has already been used in a number of scientific publications involving melt-conditioned direct-chill (MC-DC) casting [@Lebon2018; @Lebon2020;] and ultrasonic processing [@Lebon2019a; @Lebon2019b; @Subroto2021]. It has also been presented in introductory modeling doctoral courses that are provided by the Liquid Metal Engineering training centre at Brunel University.
[…]
