+++ paper/paper.bib
@@ -1,296 +1,277 @@
 @misc{acousticStreamingFoam,
-  author = {G.S.B. Lebon},
-  title = {acousticStreamingFoam: an acoustic streaming solver},
-  year = {2021},
+  author    = {G.S.B. Lebon},
+  journal   = {GitHub repository},
   publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/blebon/acousticStreamingFoam}
+  title     = {acousticStreamingFoam: an acoustic streaming solver},
+  url       = {https://github.com/blebon/acousticStreamingFoam},
+  year      = {2021}
 }
 
 @article{Andersson2002,
-  author = {Andersson, J. O. and Helander, Thomas and H{\"{o}}glund, Lars and Shi, Pingfang and Sundman, Bo},
-  doi = {10.1016/S0364-5916(02)00037-8},
-  issn = {03645916},
+  author  = {Andersson, J. O. and Helander, Thomas and H{\"{o}}glund, Lars and Shi, Pingfang and Sundman, Bo},
+  doi     = {10.1016/S0364-5916(02)00037-8},
+  issn    = {03645916},
   journal = {Calphad: Computer Coupling of Phase Diagrams and Thermochemistry},
-  number = {2},
-  title = {{Thermo-Calc & DICTRA, computational tools for materials science}},
-  volume = {26},
-  year = {2002}
+  number  = {2},
+  title   = {{Thermo-Calc & DICTRA, computational tools for materials science}},
+  volume  = {26},
+  year    = {2002}
 }
 
 @article{Bennon1987,
-  title = {A continuum model for momentum, heat and species transport in binary solid-liquid phase change systems—I. Model formulation},
+  author  = {W.D. Bennon and F.P. Incropera},
+  doi     = {https://doi.org/10.1016/0017-9310(87)90094-9},
+  issn    = {0017-9310},
   journal = {International Journal of Heat and Mass Transfer},
-  volume = {30},
-  number = {10},
-  pages = {2161-2170},
-  year = {1987},
-  issn = {0017-9310},
-  doi = {https://doi.org/10.1016/0017-9310(87)90094-9},
-  url = {https://www.sciencedirect.com/science/article/pii/0017931087900949},
-  author = {W.D. Bennon and F.P. Incropera}
+  number  = {10},
+  pages   = {2161-2170},
+  title   = {A continuum model for momentum, heat and species transport in binary solid-liquid phase change systems—I. Model formulation},
+  url     = {https://www.sciencedirect.com/science/article/pii/0017931087900949},
+  volume  = {30},
+  year    = {1987}
 }
 
 @book{Eskin2008,
-  author = {Eskin, Dmitry G.},
+  author    = {Eskin, Dmitry G.},
   booktitle = {Physical Metallurgy of Direct Chill Casting of Aluminum Alloys},
-  doi = {10.1201/9781420062823},
-  isbn = {9781420062823},
-  pages = {1--303},
+  doi       = {10.1201/9781420062823},
+  isbn      = {9781420062823},
+  pages     = {1--303},
   publisher = {CRC Press},
-  title = {{Physical metallurgy of direct chill casting of aluminum alloys}},
-  year = {2008}
+  title     = {{Physical metallurgy of direct chill casting of aluminum alloys}},
+  year      = {2008}
 }
 
 @book{Eskin2014,
-  author = {Eskin, Georgy I. and Eskin, Dmitry G.},
+  author    = {Eskin, Georgy I. and Eskin, Dmitry G.},
   booktitle = {Ultrasonic Treatment of Light Alloy Melts, Second Edition},
[…]
