+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @misc{acousticStreamingFoam,
-  author    = {G.S.B. Lebon},
+  author    = {Lebon, G.S.B.},
   journal   = {GitHub repository},
   publisher = {GitHub},
   title     = {acousticStreamingFoam: an acoustic streaming solver},
@@ -19,12 +19,12 @@
 }
 
 @article{Bennon1987,
-  author  = {W.D. Bennon and F.P. Incropera},
+  author  = {Bennon, W. D. and Incropera, F. P.},
   doi     = {https://doi.org/10.1016/0017-9310(87)90094-9},
   issn    = {0017-9310},
   journal = {International Journal of Heat and Mass Transfer},
   number  = {10},
-  pages   = {2161-2170},
+  pages   = {2161--2170},
   title   = {A continuum model for momentum, heat and species transport in binary solid-liquid phase change systems—I. Model formulation},
   url     = {https://www.sciencedirect.com/science/article/pii/0017931087900949},
   volume  = {30},
@@ -38,7 +38,7 @@
   isbn      = {9781420062823},
   pages     = {1--303},
   publisher = {CRC Press},
-  title     = {{Physical metallurgy of direct chill casting of aluminum alloys}},
+  title     = {{Physical Metallurgy of Direct Chill Casting of Aluminum Alloys}},
   year      = {2008}
 }
 
@@ -46,6 +46,8 @@
   author    = {Eskin, Georgy I. and Eskin, Dmitry G.},
   booktitle = {Ultrasonic Treatment of Light Alloy Melts, Second Edition},
   doi       = {10.1201/b17270},
+  isbn      = {9781466577985},
+  publisher = {CRC Press},
   title     = {{Ultrasonic Treatment of Light Alloy Melts, Second Edition}},
   year      = {2014}
 }
@@ -53,9 +55,11 @@
 @article{Faden2018,
   author  = {Faden, Moritz and K{\"{o}}nig-Haagen, Andreas and H{\"{o}}hlein, Stephan and Br{\"{u}}ggemann, Dieter},
   doi     = {10.1016/j.ijheatmasstransfer.2017.10.033},
-  issn    = {00179310},
+  issn    = {0017-9310},
   journal = {International Journal of Heat and Mass Transfer},
+  pages   = {757--767},
   title   = {{An implicit algorithm for melting and settling of phase change material inside macrocapsules}},
+  url     = {https://www.sciencedirect.com/science/article/pii/S0017931017329241},
   volume  = {117},
   year    = {2018}
 }
@@ -63,9 +67,11 @@
 @article{Hatic2018,
   author  = {Hati{\'{c}}, Vanja and Mavri{\v{c}}, Bo{\v{s}}tjan and Ko{\v{s}}nik, Nejc and {\v{S}}arler, Bo{\v{z}}idar},
   doi     = {10.1016/j.apm.2017.09.034},
-  issn    = {0307904X},
+  issn    = {0307-904X},
   journal = {Applied Mathematical Modelling},
+  pages   = {170--188},
   title   = {{Simulation of direct chill casting under the influence of a low-frequency electromagnetic field}},
+  url     = {https://www.sciencedirect.com/science/article/pii/S0307904X17305863},
   volume  = {54},
   year    = {2018}
 }
@@ -97,29 +103,30 @@
   issn    = {19961944},
   journal = {Materials},
   number  = {19},
+  pages   = {3262},
   title   = {{Numerical modelling of the ultrasonic treatment of aluminium melts: An overview of recent advances}},
   volume  = {12},
   year    = {2019}
 }
 
 @article{Lebon2019b,
[…]
