+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @misc{acousticStreamingFoam,
-  author    = {Lebon, G.S.B.},
+  author    = {Lebon, Bruno},
   journal   = {GitHub repository},
   publisher = {GitHub},
   title     = {acousticStreamingFoam: an acoustic streaming solver},
@@ -90,7 +90,7 @@
 }
 
 @inproceedings{Lebon2018,
-  author    = {Lebon, G S Bruno and Li, Hu-Tian and Patel, Jayesh B and Assadi, H and Fan, Z},
+  author    = {Lebon, Bruno and Li, Hu-Tian and Patel, Jayesh B. and Assadi, H and Fan, Zhongyun},
   booktitle = {Magnesium 2018},
   pages     = {100--106},
   title     = {{Numerical modelling of melt conditioned direct-chill (MC-DC) casting of AZ31 magnesium alloy}},
@@ -110,7 +110,7 @@
 }
 
 @article{Lebon2019b,
-  author  = {Lebon, G. S.Bruno and Salloum-Abou-Jaoude, Georges and Eskin, Dmitry and Tzanakis, Iakovos and Pericleous, Koulis and Jarry, Philippe},
+  author  = {Lebon, Bruno and Salloum-Abou-Jaoude, Georges and Eskin, Dmitry and Tzanakis, Iakovos and Pericleous, Koulis and Jarry, Philippe},
   doi     = {10.1016/j.ultsonch.2019.02.002},
   issn    = {18732828},
   journal = {Ultrasonics Sonochemistry},
@@ -124,7 +124,7 @@
 }
 
 @article{Lebon2020,
-  author   = {Lebon, G. S. Bruno and Li, Hu Tian and Patel, Jayesh B. and Assadi, Hamid and Fan, Zhongyun},
+  author   = {Lebon, Bruno and Li, Hu Tian and Patel, Jayesh B. and Assadi, Hamid and Fan, Zhongyun},
   doi      = {10.1016/j.apm.2019.08.032},
   issn     = {0307-904X},
   journal  = {Applied Mathematical Modelling},
@@ -151,7 +151,7 @@
 }
 
 @article{Ni1991,
-  author  = {Ni, J and Beckermann, C},
+  author  = {Ni, J. and Beckermann, C.},
   doi     = {10.1007/BF02651234},
   issn    = {1543-1916},
   journal = {Metallurgical Transactions B},
@@ -187,7 +187,7 @@
 }
 
 @article{Sree2021,
-  author  = {Sree, Sree Manu and Barekar, Nilam S. and Lazaro-Nebreda, Jaime and Patel, Jayesh B. and Fan, Zhongyun},
+  author  = {{Sree Manu}, K. M. and Barekar, Nilam S. and Lazaro-Nebreda, Jaime and Patel, Jayesh B. and Fan, Zhongyun},
   doi     = {10.1016/j.jmatprotec.2021.117170},
   issn    = {0924-0136},
   journal = {Journal of Materials Processing Technology},
@@ -198,7 +198,7 @@
 }
 
 @article{Subroto2021,
-  author  = {Subroto, Tungky and Lebon, Gerard S.B. and Eskin, Dmitry G. and Skalicky, Ivan and Roberts, Dan and Tzanakis, Iakovos and Pericleous, Koulis},
+  author  = {Subroto, Tungky and Lebon, Bruno and Eskin, Dmitry G. and Skalicky, Ivan and Roberts, Dan and Tzanakis, Iakovos and Pericleous, Koulis},
   doi     = {10.1016/j.jmrt.2021.03.061},
   issn    = {22387854},
   journal = {Journal of Materials Research and Technology},
[…]
