### 2023-01-26 #4871 @chennachaos
Hi @blebon, I think this library is a great add-on to OpenFOAM. The functionality is described well. However, at the moment, it lacks in content in terms of documentation.

* One tutorial is not sufficient. Additional tutorials, preferably one 3D test case, need to be added to demonstrate the features further.
* Update the existing tutorial with grid plots with appropriate annotations. Also, add the details of runtime at the execution stage.
* It is not quite clear what the Python helper script does. Add some details on this. Also, it is not clear how to set up different problems. Additional details on this would be useful for the users interested in using this library.
* The library lacks automated tests to verity/test its functionality.
* Community guidelines on contributing to the library and reporting issues are missing.
