### 2023-01-26 1156082
JOSS review #2: Contributing and bug reports guidelines
