+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -24,8 +24,9 @@ A clear and concise description of what you expected to happen.
 If applicable, add screenshots, contour/vector plots, graphs, or any other visual aid to help explain your problem.
 
 **Desktop (please complete the following information):**
- - OS: [e.g. Ubuntu]
- - Version [e.g. OF9]
+
+- OS: [e.g. Ubuntu]
+- Version [e.g. OF9]
 
 **Additional context**
 Add any other context about the problem here.
