+++ .github/PULL_REQUEST_TEMPLATE/pull_request_template.md
@@ -1,13 +1,16 @@
+# Pull Request for directChillFoam
 
 👋👋 Thank you for wanting to contribute to directChillFoam. 👋👋
 
-* [ ] Before submitting, please read the [How to Contribute to OpenFOAM guide](https://openfoam.org/dev/how-to-contribute/). Adhere to OpenFOAM's coding convention.
-* [ ] Does your contribution pass the tutorial validation (predicted temperature matches experimental data)?
+## **Contributor checklist**
 
-#### **Did you write a bug fix patch?**
+* [ ] Before submitting, please read the [How to Contribute to OpenFOAM guide](https://openfoam.org/dev/how-to-contribute/). Adhere to OpenFOAM's coding convention.
+* [ ] Does your contribution pass the tutorials validation (predicted temperature matches experimental data)?
+
+## **Did you write a bug fix patch?**
 
 * [ ] Describe the problem and the solution as clearly as possible.
 
-#### **Is you fix purely cosmetic?**
+## **Is you fix purely cosmetic?**
 
-* [ ] Cosmetic changes (spaces/tabs ...) will be rejected.
+* [ ] My change is not cosmetic. I understand that cosmetic changes (spaces/tabs ...) will be rejected.
