### 2023-01-26 108070c
JOSS review #2: Contributing and bug reports guidelines
