+++ tutorials/heatTransfer/directChillFoam/Subroto2021/plot_sump.py
@@ -1,0 +1,217 @@
+#!/usr/bin/env python3
+"""Plots the sump profile.
+
+This script plots the sump profile using the output in the
+postProcessing directory.
+
+Example:
+    Run python script in case root directory:
+
+        $ python3 plot_sump.py
+
+Todo:
+    * None
+
+.. _Google Python Style Guide:
+   http://google.github.io/styleguide/pyguide.html
+
+"""
+
+from __future__ import division, print_function
+
+from os import listdir, path, sep
+from natsort import natsorted, ns
+
+import vtk
+from vtk.numpy_interface import dataset_adapter as dsa
+from vtk.util.numpy_support import vtk_to_numpy
+
+import matplotlib.tri as tr
+import matplotlib.pyplot as plt
+
+from natsort import natsorted, ns
+from numpy import (
+    linspace,
+    sqrt,
+    around,
+    where,
+)
+
+__author__ = "Bruno Lebon"
+__copyright__ = "Copyright, 2022, Brunel University London"
+__credits__ = ["Bruno Lebon"]
+__email__ = "Bruno.Lebon@brunel.ac.uk"
+__status__ = "Production"
+
+
+def load_slice(filename):
+    """Read field from .vtp file.
+
+    :param filename: Filename of .vtp file, including path.
+    :type filename: str
+    :return: A tuple containing points, triangulation and scalar field.
+    :rtype: tuple[numpy.array]
+
+    """
+    if not path.exists(filename):
+        return None
+    reader = vtk.vtkXMLPolyDataReader()
+    reader.SetFileName(filename)
+    reader.Update()
+
+    data = reader.GetOutput()
+    arrays = dsa.WrapDataObject(data)
+
+    # Extracting triangular information
+    cells = arrays.GetPolys()
+    triangles = vtk_to_numpy(cells.GetData())
+
+    ntri = cells.GetNumberOfCells()
+    tris = [None] * ntri
+    tri = [None] * ntri
+
+    j = 0
+    offset = 0.1
+
+    for _ in range(ntri):
+        points = triangles[j]
+        tris[_] = list(triangles[j : j + points + 1])
+
+        tri[_ + offset] = list([triangles[j + 1], triangles[j + 2], triangles[j + 3]])
+
+        if points == 4:
+            offset += 1
+            tri.insert(
+                _ + offset, list([triangles[j + 3], triangles[j + 4], triangles[j + 1]])
+            )
+
+        j += points + 1
+
+    # Extracting points
+    points = vtk_to_numpy(arrays.GetPoints())
+    x, y, z = points.T
+
+    # Extract data
+    sc = vtk_to_numpy(arrays.GetPointData().GetArray(0))
+
+    if any([i in filename for i in ["U", "grad(T)"]]):
+        return (x, y, z, tri, sc.T)
+    else:
+        return (x, y, z, tri, sc)
+
+
+Ucast = 0.00233  # Casting speed in m/s
+offset = 0  # Set to mould position
+
+T_sol = 805.503  # Solidus temperature in K
+T_liq = 928.236  # Liquidus temperature in K
+
+
+def plot_sump(image_name="sump.png", cmap="jet"):
+    """Plots sump contour.
+
+    :param image_name: Filename of image, including path.
+    :type image_name: str
+    :param cmap: Color map for plotting contour.
+    :type cmap: str or matplotlib.colormaps
[…]

[…]
