### 2023-01-26 #4871 @blebon
@chennachaos, thank you for the constructive comments. I am working on your comments and hope to resolve all of them in a timely manner.

While going through them, I had a few questions regarding some of your comments:

1. "It is not quite clear what the Python helper script does.": I already have documentation for the Python scripts that are used in the tutorials: [https://blebon.com/directChillFoam/tutorials/modules.html](https://blebon.com/directChillFoam/tutorials/modules.html). I am not quite sure what to add more here. Would you like to see a long description for each function?
2. "The library lacks automated tests to verity/test its functionality.": Would a benchmark validation test like [this](https://github.com/blebon/directChillFoam/tree/master/tutorials/heatTransfer/directChillFoam/Vreeman2002/test) be adequate to address your comment?
3. The community guidelines were in the [.github folder](https://github.com/blebon/directChillFoam/tree/master/.github). Would you like to see more description in there or do you think they should have been placed somewhere else. For now, I have added links to these in the root README.md.

Thanks again for the comments. I am working on adding more tutorials, including a 3D case. I will add figures for describing the meshes that are used and a rough estimate of the run time on my test virtual environment (this will go in the  Sphinx documentation). I will also expand the documentation to help a new OpenFOAM user to setup a new case. The latter might take some time but hope to get everything sorted by early next week.
