### 2023-01-26 9e50a1b
BOOST_CHECK_CLOSE for floating points. BOOST_CHECK_CLOSE_COLLECTIONS not available.
