+++ tutorials/heatTransfer/directChillFoam/Subroto2021/system/cylinder.py
@@ -1,17 +1,17 @@
-#!/usr/bin/env python3
-"""Create axisymmetric block mesh dict file
+"""Writes a blockMeshDict file for an axisymmetric case.
+This script writes a blockMeshDict file for an axisymmetric case in the system directory.
+
+Refer to https://doc.cfd.direct/openfoam/user-guide-v6/blockmesh for the official blockMesh utility documentation.
 
 Example:
-    Run python script in system directory:
-
-        $ python3 cylinder.py
-
+    Run python script in case system directory:
+        $ cd system
+        $ python convert_vtk.py
 Todo:
     * None
-
 .. _Google Python Style Guide:
    http://google.github.io/styleguide/pyguide.html
-
+   
 """
 
 from __future__ import division, print_function
@@ -25,11 +25,18 @@ __email__ = "Bruno.Lebon@brunel.ac.uk"
 __status__ = "Production"
 
 
-def write_vertices(
-    angle=2.5, diameter=155.0, z_points=(-300.0, -57.0, -40.0, 0.0, 170.0)
-):
-    """
-    Vertices
+def write_vertices(angle=2.5, diameter=155.0, z_points=(-250.0, -50.0, 0.0, 50.0)):
+    """Writes the vertices entry
+
+    :param angle: Angle of the axisymmetric slice.
+    :type angle: float
+    :param diameter: Diameter of the cast billet (mm).
+    :type diameter: float
+    :param z_points: z coordinates of the boundaries of the billet and mould sections.
+    :type z_points: tuple
+    :return: string containing the blockMeshDict vertices entry
+    :rtype: string
+
     """
     block = "vertices\n"
     block += "(\n"
@@ -37,6 +44,7 @@ def write_vertices(
     radius = diameter / 2
     x = radius * cos(pi * angle / 180)
     y = radius * sin(pi * angle / 180)
+    # Write each vertex coordinate and label the vertex number in a comment at the right
     for z in z_points:
         block += f"    ({0:7.4f}  {0:7.4f} {z:6.1f}) // {vertex:2d}\n"
         vertex += 1
@@ -51,29 +59,47 @@ def write_vertices(
 
 
 def write_blocks(
-    z_points=(-331.0, -71.0, -57.0, -40.0, 0.0, 50.0), diameter=155, mesh=1.0
+    diameter=155.0, z_points=(-250.0, -50.0, 0.0, 50.0), rmesh=1.0, zmesh=1.0
 ):
-    """
-    Blocks
+    """Writes the blocks entry
+
+    :param diameter: Diameter of the cast billet (mm).
+    :type diameter: float
+    :param z_points: z coordinates of the boundaries of the billet and mould sections.
+    :type z_points: tuple
+    :param rmesh: Radial mesh adjustment parameter.
+    :type rmesh: float
+    :param zmesh: Vertical mesh adjustment parameter.
+    :type zmesh: float
+    :return: string containing the blockMeshDict blocks entry
+    :rtype: string
+
     """
     block = "blocks\n"
     block += "(\n"
     prism = 0
     N = 3
-    rcells = int(diameter / (2 * mesh))
+    rcells = int(diameter / (2 * rmesh))
     for i in range(len(z_points) - 1):
-        cells = int((z_points[i + 1] - z_points[i]) / mesh)
+        cells = int((z_points[i + 1] - z_points[i]) / zmesh)
         block += f"    hex ({0+N*i:2d} {2+N*i:2d} {1+N*i:2d} {0+N*i:2d} {3+N*i:2d} {5+N*i:2d} {4+N*i:2d} {3+N*i:2d}) domain ({rcells:2d} 1 {cells:3d}) simpleGrading (0.5 1 1) // {prism:2d}\n"
         prism += 1
     block += ");\n"
     return block
 
 
[…]
