### 2023-01-27 f506f30
JOSS review #2: Added execution time estimates to tutorials.
