### 2023-01-27 7d134cb
JOSS review #2: Added 3D tutorial case.
