+++ tutorials/heatTransfer/directChillFoam/Lebon2020/system/cylinder.py
@@ -1,0 +1,368 @@
+"""Writes a blockMeshDict file for an axisymmetric case.
+This script writes a blockMeshDict file for an axisymmetric case in the system directory.
+
+Refer to https://doc.cfd.direct/openfoam/user-guide-v6/blockmesh for the official blockMesh utility documentation.
+
+Example:
+    Run python script in case system directory:
+        $ cd system
+        $ python convert_vtk.py
+Todo:
+    * None
+.. _Google Python Style Guide:
+   http://google.github.io/styleguide/pyguide.html
+   
+"""
+
+
+from numpy import pi, cos
+
+__author__ = "Bruno Lebon"
+__copyright__ = "Copyright, 2023, Brunel University London"
+__credits__ = ["Bruno Lebon"]
+__email__ = "Bruno.Lebon@brunel.ac.uk"
+__status__ = "Production"
+
+
+def write_vertices(
+    square=16, diameter=155.0, z_points=(-300.0, -57.0, -40.0, 0.0, 170.0)
+):
+    """Writes the vertices entry
+
+    :param square: Length of inner square section of mesh.
+    :type angle: float
+    :param diameter: Diameter of the cast billet (mm).
+    :type diameter: float
+    :param z_points: z coordinates of the boundaries of the billet and mould sections.
+    :type z_points: tuple
+    :return: string containing the blockMeshDict vertices entry
+    :rtype: string
+
+    """
+    block = "vertices\n"
+    block += "(\n"
+    vertex = 0
+    curve = (diameter / 2) * cos(45.0 * pi / 180.0)
+    # Write each vertex coordinate and label the vertex number in a comment at the right
+    for z in z_points:
+        block += f"    ( {square:16.13f} -{square:16.13f} {z:5.1f}) // {vertex:2d}\n"
+        vertex += 1
+        block += f"    (-{square:16.13f} -{square:16.13f} {z:5.1f}) // {vertex:2d}\n"
+        vertex += 1
+        block += f"    (-{square:16.13f}  {square:16.13f} {z:5.1f}) // {vertex:2d}\n"
+        vertex += 1
+        block += f"    ( {square:16.13f}  {square:16.13f} {z:5.1f}) // {vertex:2d}\n"
+        vertex += 1
+
+        block += "    \n"
+
+        block += f"    ( {curve:16.13f} -{curve:16.13f} {z:5.1f}) // {vertex:2d}\n"
+        vertex += 1
+        block += f"    (-{curve:16.13f} -{curve:16.13f} {z:5.1f}) // {vertex:2d}\n"
+        vertex += 1
+        block += f"    (-{curve:16.13f}  {curve:16.13f} {z:5.1f}) // {vertex:2d}\n"
+        vertex += 1
+        block += f"    ( {curve:16.13f}  {curve:16.13f} {z:5.1f}) // {vertex:2d}\n"
+        vertex += 1
+
+        if z != z_points[-1]:
+            block += "    \n"
+    block += ");\n"
+    return block
+
+
+def write_blocks(z_points=(-300.0, -57.0, -40.0, 0.0, 170.0), zmesh=1.0):
+    """Writes the blocks entry
+
+    :param z_points: z coordinates of the boundaries of the billet and mould sections.
+    :type z_points: tuple
+    :param zmesh: Vertical mesh adjustment parameter.
+    :type zmesh: float
+    :return: string containing the blockMeshDict blocks entry
+    :rtype: string
+
+    """
+    block = "blocks\n"
+    block += "(\n"
+    vertex = 0
[…]
