+++ tutorials/heatTransfer/directChillFoam/Lebon2020/plot_line.py
@@ -1,0 +1,157 @@
+#!/usr/bin/env python3
+"""Plots the temperature profiles at prescribed locations.
+
+This script plots the comparison between numerical and experimental
+data. Experimental data in system/validation.dat. Numerical data in
+postProcessing directory.
+
+Example:
+    Run python script in case root directory:
+
+        $ python3 plot_line.py
+
+Todo:
+    * None
+
+.. _Google Python Style Guide:
+   http://google.github.io/styleguide/pyguide.html
+
+"""
+
+from __future__ import division, print_function
+
+from os import listdir
+
+# from time import sleep
+
+import matplotlib.pyplot as plt
+from natsort import natsorted
+from numpy import isfinite
+from pandas import read_csv, merge
+
+__author__ = "Bruno Lebon"
+__copyright__ = "Copyright, 2023, Brunel University London"
+__credits__ = ["Bruno Lebon"]
+__email__ = "Bruno.Lebon@brunel.ac.uk"
+__status__ = "Production"
+
+
+def read_numerical_dataframe(
+    plot_time,
+    probe_position="centreline",
+    temperature_column_name="tcK",
+):
+    """Read temperature at plot_time and return as dataframe."""
+    _df = read_csv(
+        f"postProcessing/{probe_position}/{plot_time:s}/line_T.xy",
+        header=None,
+        delimiter=r"\s+",
+    )
+    _df[0] *= -1  # z inverted in experiment
+    _df.columns = ["z", temperature_column_name]
+    return _df
+
+
+def plot_line(plot_time, Experimental, Image_Path=".", offset=0.0):
+    """Plots temperature graphs at plot_time.
+
+    :param plot_time: Time at which to plot the temperatures.
+    :type plot_time: str
+    :param Experimental: Dataframe containing experimental data.
+    :type Experimental: pandas.Dataframe
+    :param Image_Path: Absolute path of line plot.
+    :type Image_Path: str
+    :param offset: Offset position of mould bottom in experiment.
+    :type offset: float
+    :return: 0 if successful.
+    :rtype: int
+
+    """
+    _plot_time_float = float(plot_time)
+    print(f"{_plot_time_float:4.1f}")
+
+    # Get numerical temperatures
+    Num_x0mm = read_numerical_dataframe(
+        plot_time, probe_position="x0mm", temperature_column_name="tK0mm"
+    )
+    Num_28mm = read_numerical_dataframe(
+        plot_time, probe_position="x28mm", temperature_column_name="tK28mm"
+    )
+    Num_44mm = read_numerical_dataframe(
+        plot_time, probe_position="x44mm", temperature_column_name="tK44mm"
+    )
+    Num_75mm = read_numerical_dataframe(
+        plot_time, probe_position="x75mm", temperature_column_name="tK75mm"
+    )
+
+    # Merge dataframes
+    Numerical = merge(Num_x0mm, Num_28mm, on="z")
+    Numerical = merge(Numerical, Num_44mm, on="z")
+    Numerical = merge(Numerical, Num_75mm, on="z")
+
+    # Convert to degrees Celcius
+    Numerical["T0mm"] = Numerical["tK0mm"] - 273.15
+    Numerical["T28mm"] = Numerical["tK28mm"] - 273.15
+    Numerical["T44mm"] = Numerical["tK44mm"] - 273.15
+    Numerical["T75mm"] = Numerical["tK75mm"] - 273.15
+
+    # Plot experimental values, masking NaNs
[…]

[…]
