### 2023-01-27 f12613e
JOSS review #2: Added 3D tutorial case.
