### 2023-01-27 d980621
JOSS review #2: 3D case tutorial
