+++ docs/source/tutorials/modules.rst
@@ -41,6 +41,45 @@ Sump contour
 .. automodule:: Vreeman2002.plot_sump
    :members:
 
+Lebon *et al.* (2020)
+=======================
+
+blockMeshDict writer
+--------------------
+
+.. automodule:: Lebon2020.system.cylinder
+   :members:
+
+VTK conversion
+--------------
+
+.. automodule:: Lebon2020.convert_vtk
+   :members:
+
+PVD generation
+--------------
+
+.. automodule:: Lebon2020.generate_pvd
+   :members:
+
+HTC file creation
+-----------------
+
+.. automodule:: Lebon2020.write_htc
+   :members:
+
+Temperature validation
+----------------------
+
+.. automodule:: Lebon2020.plot_line
+   :members:
+
+Sump contour
+------------
+
+.. automodule:: Lebon2020.plot_sump
+   :members:
+
 Subroto *et al.* (2021)
 =======================
 
