+++ tutorials/heatTransfer/directChillFoam/Lebon2020/test/README.md
@@ -1,0 +1,15 @@
+# Testing the tutorial case
+
+Test if the correct temperatures are predicted.
+Requires cmake and the boost C++ libraries.
+
+After running the tutorial:
+
+1. Create an empty build directory:
+    ``mkdir build``
+2. Enter the build directory and run cmake:
+    ``cd build; cmake ..``
+3. Compile the test:
+    ``make``
+4. Run the test:
+    ``./test_numerical``

[…]
