### 2023-01-29 #4871 @chennachaos
@blebon,
1. I think this is good enough. You may consider pointing to this page next when the Python script is used.
2. This test looks like an integration test. It's better to have some unit tests to test the key functions.
3. The main README.md file or the documentation would be a better place for the community guidelines.

I hope that clarifies.
