### 2023-01-30 874b505
JOSS review #2: Expanded 3D tutorial explanations.
