+++ docs/source/conf.py
@@ -50,6 +50,9 @@ extensions = [
 ]
 autosectionlabel_prefix_document = True
 
+# Add figures numbering
+numfig = True
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 
