+++ docs/source/solver/directChillFoam.rst
@@ -1,5 +1,9 @@
+================================
 Direct-chill (DC) casting solver
 ================================
+
+.. contents:: Contents:
+  :backlinks: none
 
 directChillFoam is based on `buoyantPimpleFoam <https://github.com/OpenFOAM/OpenFOAM-9/tree/master/applications/solvers/heatTransfer/buoyantPimpleFoam>`_, a transient solver for buoyant, turbulent flow of compressible fluids and heat-transfer. directChillFoam adds the following capabilities to the solver:  
 
@@ -11,7 +15,7 @@ directChillFoam is based on `buoyantPimpleFoam <https://github.com/OpenFOAM/Open
 Definition in file directChillFoam.C
 
 Installation
-------------
+============
 
 Pre-requisites:  
 
@@ -31,7 +35,7 @@ In the directChillFoam/applications/solver/heatTransfer/directChillFoam director
   $ wmake
 
 Running the application
------------------------
+=======================
 
 In the case directory, run:
 
@@ -40,12 +44,12 @@ In the case directory, run:
   $ directChillFoam
 
 Theory
-------
+======
 
 Full details of the theory behind the DC casting solver implementation can be found in G.S. Bruno Lebon, Georges Salloum-Abou-Jaoude, Dmitry Eskin, Iakovos Tzanakis, Koulis Pericleous, Philippe Jarry, "Numerical modelling of acoustic streaming during the ultrasonic melt treatment of direct-chill (DC) casting", *Ultrasonics Sonochemistry* **54** (2019) 171-182 `doi:10.1016/j.ultsonch.2019.02.002 <https://doi.org/10.1016/j.ultsonch.2019.02.002>`_.
 
 Nomenclature
-------------
+============
 
 .. table:: Variables used in direct chill casting simulations.
   :widths: auto
