+++ docs/source/src/ThermophysicalTransportModels.rst
@@ -1,12 +1,16 @@
+===================
 Boundary conditions
 ===================
+
+.. contents:: Contents:
+  :backlinks: none
 
 Robin boundary conditions apply a heat flux condition to temperature on an external wall.
 
 Two boundary conditions are available in this library: the `mould HTC <#mould-heat-transfer-coefficient>`_ and the `water film HTC <#water-film-heat-transfer-coefficient>`_.
 
 Installation
-------------
+============
 
 Pre-requisites:  
 
@@ -19,13 +23,13 @@ In the directChillFoam/src/ThermophysicalTransportModels directory, run:
   $ wmake libso
 
 Mould heat transfer coefficient
--------------------------------
+===============================
 
 .. doxygenclass:: Foam::mouldHTCFvPatchScalarField
   :members:
 
 Water film heat transfer coefficient
-------------------------------------
+===============================
 
 .. doxygenclass:: Foam::waterFilmHTCFvPatchScalarField
   :members:

[…]
