+++ docs/source/tutorials/Subroto2021.rst
@@ -5,7 +5,8 @@ Tutorial: Subroto *et al.* (2021) - DC casting of a ternary Al alloy
 .. contents:: Contents:
   :backlinks: none
 
-This tutorial describes how to pre-process, run and post-process a direct chill (DC) casting case with directChillFoam.
+This tutorial describes the steps to pre-process, run and post-process a 3D direct chill (DC) casting case that corresponds to the AA6008 conventional DC casting experiment described in `Subroto et al. (2021) <https://doi.org/10.1016/j.jmrt.2021.03.061>`_ with directChillFoam.
+
 
 .. figure:: ../images/DC_Schematic.png
   :width: 80%
@@ -20,13 +21,17 @@ Pre-processing
 
 The schematic for the DC casting setup of this case is as follows:
 
+.. _Subroto2021_BC:
+
 .. Figure:: ../images/Subroto2021_BC.png
   :width: 80%
   :alt: Schematic of the Subroto *et al.* (2021) case
 
   Boundary patch labels and dimensions of the casting mould and billet. The image is not drawn to scale.
 
-Change to the case directory.
+The first step is to generate the mesh for the case and setup the boundary conditions.
+
+To run the pre-processing steps, change to the case directory.
 
 .. code-block:: console
 
@@ -35,7 +40,7 @@ Change to the case directory.
 Mesh generation
 ---------------
 
-Move into the system directory.
+The dictionary file that describes the mesh is located in the system directory.
 
 .. code-block:: console
 
@@ -47,7 +52,9 @@ Generate the blockMeshDict file using ``Subroto2021.system.cylinder.write_blockM
 
   $ python3 cylinder.py
 
-Return to the case directory and generate the mesh.
+The axisymmetryc mesh consists of a slice of a cylinder that matches the billet dimensions. The z_points tuple list the vertices z coordinates as defined in :numref:`lebon2020_bc`. The patch names are entered in the faces list in the ``Lebon2020.system.cylinder.write_boundary()`` function.
+
+Return to the case directory and generate the mesh using `blockMesh <https://doc.cfd.direct/openfoam/user-guide-v6/blockmesh>`_.
 
 .. code-block:: console
 
@@ -57,11 +64,14 @@ Return to the case directory and generate the mesh.
 Boundary and initial conditions
 -------------------------------
 
-Initialize the fields in the case directory:
+To speed up the simulation, initialize the fields with rough estimates of temperatures (and corresponding liquid fractions) in the case directory:
 
 .. code-block:: console
 
   $ setFields
+
+.. warning::
+  The liquid fraction values must match the temperatures for the alloy according to the constant/tStar Foam table. Otherwise, the case can diverge.
 
 Solute
 ^^^^^^
@@ -82,7 +92,7 @@ The fields are initialized with constant values. Use 0.6 for silicon and 0.48 fo
 
   internalField   uniform 0.48; // C.Mg
 
-The inlet boundary condition is a fixed value:
+The inlet boundary condition is a fixed value (stored in internalField):
 
 .. code-block:: C 
   
@@ -178,6 +188,8 @@ The casting velocity :math:`{U}` is prescribed at the mould and water-cooled sur
       #};
     }
 
+The values are initialized at the casting speed in m s\ :sup:`-1`.
+
 Solid velocity
 ^^^^^^^^^^^^^^
 
@@ -191,29 +203,31 @@ The movingShell coded function from the liquid velocity boundary condition is re
       value           $internalField;
   }
 
+The values are also initialized at the casting speed in m s\ :sup:`-1`.
+
 Physical properties
 -------------------
 
 Transport properties
 ^^^^^^^^^^^^^^^^^^^^
 
[…]

[…]
