+++ tests/fvConstraints/case/0/C.Cu
@@ -1,0 +1,46 @@
+/*--------------------------------*- C++ -*----------------------------------*\
+  =========                 |
+  \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
+   \\    /   O peration     | Website:  https://openfoam.org
+    \\  /    A nd           | Version:  9
+     \\/     M anipulation  |
+\*---------------------------------------------------------------------------*/
+FoamFile
+{
+    format      ascii;
+    class       volScalarField;
+    location    "0";
+    object      C.Cu;
+}
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
+
+dimensions      [0 0 0 0 0 0 0];
+
+internalField   uniform 0.06;
+
+boundaryField
+{
+    "(water-film|mould|air-gap|graphite|ceramic|hot-top)"
+    {
+        type            zeroGradient;
+    }
+
+    free-surface
+    {
+        type            fixedValue;
+        value           $internalField;
+    }
+
+    ram
+    {
+        type            zeroGradient;
+    }
+
+    symmetry_planes
+    {
+        type            symmetry;
+    }
+}
+
+
+// ************************************************************************* //
