+++ tests/fvConstraints/test_limitSolute.C
@@ -18,7 +18,6 @@ Author
 #include "multicomponentAlloy.H"
 #include "fluidThermo.H"
 #include "fvConstraints.H"
-// #include "localEulerDdtScheme.H"
 
 namespace utf = boost::unit_test;
 
@@ -30,12 +29,7 @@ struct F
         : argc(utf::framework::master_test_suite().argc),
           argv(utf::framework::master_test_suite().argv)
     {
-        #include "setRootCaseLists.H"
-        #include "createTime.H"
-        #include "createDynamicFvMesh.H"
-        #include "createFields.H"
-
-        Info << "\nStarting limitSolute tests\n" << endl;
+        BOOST_TEST_MESSAGE("\nStarting limitSolute tests\n");
     }
 
     ~F()
@@ -43,6 +37,7 @@ struct F
         Info << "\nEnd\n" << endl;
     }
 
+    public:
     int argc;
     char **argv;
 };
@@ -51,7 +46,17 @@ BOOST_FIXTURE_TEST_SUITE(CheckLimitSoluteFvConstraint, F);
 
     BOOST_AUTO_TEST_CASE(CheckIfLimitSoluteFvConstraintHasBeenRead)
     {       
-        Info << "CheckIfLimitSoluteFvConstraintHasBeenRead" << endl;
+        #include "setRootCaseLists.H"
+        #include "createTime.H"
+        #include "createDynamicFvMesh.H"
+        #include "createFields.H"
+        
+        BOOST_TEST_MESSAGE("-- Checking if a limitSolute dictionary entry has been read");
+        BOOST_WARN_EQUAL(fvConstraints.PtrListDictionary<fvConstraint>::size(), 1);
+
+        BOOST_TEST_MESSAGE("-- Checking if C.Cu limitSolute dictionary has been read");
+        BOOST_REQUIRE_EQUAL(fvConstraints.constrainsField("C.Cu"), 1);
+        BOOST_REQUIRE_EQUAL(fvConstraints.constrainsField("C.Si"), 0);
     }
 
 BOOST_AUTO_TEST_SUITE_END();
