### 2023-01-31 b20a38c
JOSS review #2: Unit test for soluteModel.
