+++ tests/soluteModel/createFields.H
@@ -1,0 +1,88 @@
+Info<< "Reading thermophysical properties\n" << endl;
+
+autoPtr<fluidThermo> pThermo(fluidThermo::New(mesh));
+fluidThermo& thermo = pThermo();
+thermo.validate(args.executable(), "h", "e");
+
+// Reading transport properties for Us 
+// evaluation
+IOdictionary transportProperties
+(
+    IOobject
+    (
+        "transportProperties",
+        runTime.constant(),
+        mesh,
+        IOobject::MUST_READ_IF_MODIFIED,
+        IOobject::NO_WRITE
+    )
+);
+
+// Liquid density
+dimensionedScalar rho2
+(
+    "rho2",
+    dimDensity,
+    transportProperties.lookup("rho2")
+);
+
+volScalarField rho
+(
+    IOobject
+    (
+        "rho",
+        runTime.timeName(),
+        mesh,
+        IOobject::NO_READ,
+        IOobject::NO_WRITE
+    ),
+    thermo.rho()
+);
+
+Info<< "Reading field U\n" << endl;
+volVectorField U
+(
+    IOobject
+    (
+        "U",
+        runTime.timeName(),
+        mesh,
+        IOobject::MUST_READ,
+        IOobject::AUTO_WRITE
+    ),
+    mesh
+);
+
+// Casting velocity below coherency
+// and grain velocity elsewhere
+Info<< "Reading field Us\n" << endl;
+volVectorField Us
+(
+    IOobject
+    (
+        "Us",
+        runTime.timeName(),
+        mesh,
+        IOobject::MUST_READ,
+        IOobject::AUTO_WRITE
+    ),
+    mesh
+);
+
+volScalarField alpha
+(
+    IOobject
+    (
+        "alpha",
+        runTime.timeName(),
+        mesh
+    ),
+    mesh,
+    dimensionedScalar("alpha", dimless, 0)
+);
+
+#include "compressibleCreatePhi.H"
+#include "createFvModels.H"
+#include "createFvConstraints.H"
+
+multicomponentAlloy alloy("alloy", U, rho, rho2, phi, fvModels, fvConstraints);
