+++ docs/source/src/fvConstraints.rst
@@ -25,3 +25,14 @@ C++ Classes
 
 .. doxygenclass:: Foam::fv::limitSolute
   :members:
+
+Testing
+=======
+
+The unit tests for the fvConstraints libraries require the Boost libraries (>= 1.69.0). The tests are run in the tests/fvConstraints/case folder:
+
+.. code-block:: console
+  
+  $ wmake .. # Build the test executable
+  $ blockMesh # Create the mesh for the test case
+  $ ./test_limitSolute --log_level=all # Run the tests
