+++ docs/source/src/ThermophysicalTransportModels.rst
@@ -33,3 +33,14 @@ Water film heat transfer coefficient
 
 .. doxygenclass:: Foam::waterFilmHTCFvPatchScalarField
   :members:
+
+Testing
+=======
+
+The unit tests for the ThermophysicalTransportModels libraries require the Boost libraries (>= 1.69.0). The tests are run in the tests/ThermophysicalTransportModels/<boundaryType>/case folder:
+
+.. code-block:: console
+  
+  $ wmake .. # Build the test executable
+  $ blockMesh # Create the mesh for the test case
+  $ ./test_mouldHTC --log_level=all # Run the tests. Use ./test_waterFilmHTC for the waterFilmHTC boundary condition.
