+++ .github/workflows/wmake.yml
@@ -9,6 +9,7 @@ on:
       - 'src/fvConstraints/**'
       - 'src/fvModels/**'
       - 'src/ThermophysicalTransportModels/**'
+      - 'tests/**'
   pull_request:
     branches:
       - master
@@ -17,6 +18,7 @@ on:
       - 'src/fvConstraints/**'
       - 'src/fvModels/**'
       - 'src/ThermophysicalTransportModels/**'
+      - 'tests/**'
   workflow_dispatch:
 
 jobs:
@@ -52,6 +54,7 @@ jobs:
     #     sudo add-apt-repository http://dl.openfoam.org/ubuntu
     #     sudo apt-get update
     #     sudo apt-get -y install openfoam9
+    #     sudo apt-get -y install cmake libboost-all-dev
 
     - name: Install OpenFOAM 9
       shell: bash
@@ -60,6 +63,7 @@ jobs:
         sudo add-apt-repository http://dl.openfoam.org/ubuntu
         sudo apt-get update
         sudo apt-get -y install openfoam9
+        sudo apt-get -y install cmake libboost-all-dev
     
     - name: wmake multicomponentAlloy
       shell: bash
@@ -112,9 +116,11 @@ jobs:
     - name: Install testing dependencies
       run: pip install pytest pytest-md pytest-emoji
     
-    - name: Run tests
+    - name: Run unit tests
       shell: bash
+      working-directory: tests
       continue-on-error: true
       run: |
         source /opt/openfoam9/etc/bashrc || true
         pytest
+        ./Alltest
