### 2023-01-31 24ac0d0
Soft link to applications and src folders in /user/mxam/mxstbbl/OpenFOAM/mxstbbl-9
