+++ docs/source/tutorials/Subroto2021.rst
@@ -46,7 +46,7 @@ The dictionary file that describes the mesh is located in the system directory.
 
   $ cd system
 
-Generate the blockMeshDict file using ``Subroto2021.system.cylinder.write_blockMeshDict()`` function from the system/cylinder.py script.
+Generate the blockMeshDict file using ``Subroto2021.system.cylinder.write_blockMeshDict()`` function from the `system/cylinder.py script <modules.html#id18>`_.
 
 .. code-block:: console
 
