### 2023-01-31 ffb90d1
CMake: All libraries and application
