+++ CMakeLists.txt
@@ -5,16 +5,28 @@ project(directChillFoam LANGUAGES CXX C)
 # Display OpenFOAM version
 MESSAGE(STATUS "Using OpenFOAM $ENV{WM_PROJECT_VERSION}")
 
+# Check if required libraries are available
+find_library(OpenFOAM_LIB
+    NAMES OpenFOAM libOpenFOAM
+    PATHS ENV FOAM_LIBBIN
+    REGISTRY_VIEW $ENV{WM_ARCH_OPTION}
+    REQUIRED
+)
+find_library(DL_LIB 
+    NAMES dl libdl
+    REGISTRY_VIEW $ENV{WM_ARCH_OPTION}
+    REQUIRED
+)
+find_library(M_LIB
+    NAMES m libm
+    REGISTRY_VIEW $ENV{WM_ARCH_OPTION}
+    REQUIRED
+)
+
+# OpenFOAM compilation options
 set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
-
-# Check if required libraries are available
-find_library(OpenFOAM REGISTRY_VIEW $ENV{WM_ARCH_OPTION} REQUIRED)
-find_library(dl REGISTRY_VIEW $ENV{WM_ARCH_OPTION} REQUIRED)
-find_library(m REGISTRY_VIEW $ENV{WM_ARCH_OPTION} REQUIRED)
-
-# OpenFOAM compilation options
 add_compile_options(-m$ENV{WM_ARCH_OPTION})
 add_compile_definitions($ENV{WM_ARCH})
 add_compile_definitions(WM_ARCH_OPTION=$ENV{WM_ARCH_OPTION})
@@ -49,7 +61,6 @@ link_directories($ENV{FOAM_LIBBIN})
 link_directories($ENV{FOAM_USER_LIBBIN})
 
 # libmulticomponentAlloy
-
 add_library(multicomponentAlloy SHARED ${MULTICOMPONENTALLOY_DIR}/soluteModel/soluteModel.C ${MULTICOMPONENTALLOY_DIR}/multicomponentAlloy.C)
 
 target_link_libraries(multicomponentAlloy LINK_PUBLIC finiteVolume dynamicFvMesh topoChangerFvMesh meshTools sampling fvModels fvConstraints OpenFOAM dl m)
@@ -57,10 +68,78 @@ target_link_options(multicomponentAlloy PUBLIC -fuse-ld=bfd)
 target_link_options(multicomponentAlloy PUBLIC LINKER:--add-needed,--no-as-needed)
 
 # Copy libmulticomponentAlloy.so library upon successful compilation
-# add_custom_command(TARGET libmulticomponentAlloy
-#                    COMMENT "Copying libmulticomponentAlloy.so to $ENV{FOAM_USER_LIBBIN}"
-#                    POST_BUILD COMMAND cp libmulticomponentAlloy $ENV{FOAM_USER_LIBBIN}
-# )
+add_custom_command(TARGET multicomponentAlloy
+                   COMMENT "Copying libmulticomponentAlloy.so to $ENV{FOAM_USER_LIBBIN}"
+                   POST_BUILD COMMAND cp libmulticomponentAlloy.so $ENV{FOAM_USER_LIBBIN}
+)
+
+# libmyFvConstraints
+set(LIMITSOLUTE_DIR src/fvConstraints/limitSolute)
+add_library(myFvConstraints SHARED ${LIMITSOLUTE_DIR}/limitSolute.C)
+
+target_include_directories(myFvConstraints PUBLIC $ENV{FOAM_SRC}/fvConstraints/lnInclude)
[…]
