### 2023-02-01 128d27e
CMake macro and testing
