+++ CMakeLists.txt
@@ -2,28 +2,52 @@ cmake_minimum_required(VERSION 3.17 FATAL_ERROR)
 
 project(directChillFoam LANGUAGES CXX C)
 
-# Display OpenFOAM version
-MESSAGE(STATUS "Using OpenFOAM $ENV{WM_PROJECT_VERSION}")
+message(STATUS "Using OpenFOAM $ENV{WM_PROJECT_VERSION}")
 
-# Check if required libraries are available
+message(STATUS "Checking if required libraries are available")
+message(STATUS "Finding OpenFOAM library")
 find_library(OpenFOAM_LIB
     NAMES OpenFOAM libOpenFOAM
     PATHS ENV FOAM_LIBBIN
     REGISTRY_VIEW $ENV{WM_ARCH_OPTION}
     REQUIRED
 )
+
+message(STATUS "Finding dl library")
 find_library(DL_LIB 
     NAMES dl libdl
     REGISTRY_VIEW $ENV{WM_ARCH_OPTION}
     REQUIRED
 )
+
+message(STATUS "Finding m library")
 find_library(M_LIB
     NAMES m libm
     REGISTRY_VIEW $ENV{WM_ARCH_OPTION}
     REQUIRED
 )
 
-# OpenFOAM compilation options
+message(STATUS "Finding Boost library")
+if(UNIX)
+  message("-- Determining id of the Linux distributor")
+  execute_process(COMMAND lsb_release --short --id OUTPUT_VARIABLE LINUX_DISTRO OUTPUT_STRIP_TRAILING_WHITESPACE)
+  # execute_process(COMMAND lsb_release --short --release OUTPUT_VARIABLE DISTRO_VERSION OUTPUT_STRIP_TRAILING_WHITESPACE)
+  if (LINUX_DISTRO MATCHES "CentOS")
+    message("-- Using boost 1.69 on CentOS")
+    set(Boost_NO_SYSTEM_PATHS TRUE)
+    set(BOOST_ROOT "/usr/include/boost169")
+    set(BOOST_INCLUDE_DIR "/usr/include/boost169/boost")
+    set(BOOST_LIBRARYDIR "/usr/lib64/boost169")
+  endif(LINUX_DISTRO MATCHES "CentOS")
+endif(UNIX)
+find_package(Boost 1.69.0 
+    REQUIRED
+    COMPONENTS unit_test_framework
+)
+
+enable_testing()
+
+message(STATUS "Setting generic OpenFOAM compilation options")
 set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
@@ -35,88 +59,158 @@ add_compile_definitions(WM_LABEL_SIZE=$ENV{WM_LABEL_SIZE})
 add_compile_options(-Wall -Wextra -Wold-style-cast -Wnon-virtual-dtor -Wno-unused-parameter -Wno-invalid-offsetof -Wno-attributes)
 add_compile_options(-O3)
 add_compile_definitions(NoRepository)
-# -ftemplate-depth-100
 if(CMAKE_COMPILER_IS_GNUCXX)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -ftemplate-depth-100")
 endif(CMAKE_COMPILER_IS_GNUCXX)
-# -fPIC
-set(CMAKE_POSITION_INDEPENDENT_CODE ON)
+set(CMAKE_POSITION_INDEPENDENT_CODE ON) # -fPIC
 
-# Match path configuration of Make/options for directChillFoam
[…]
