### 2023-02-01 9a3ab3c
CMake workflow with tests
