+++ CMakeLists.txt
@@ -56,7 +56,9 @@ add_compile_definitions($ENV{WM_ARCH})
 add_compile_definitions(WM_ARCH_OPTION=$ENV{WM_ARCH_OPTION})
 add_compile_definitions(WM_DP)
 add_compile_definitions(WM_LABEL_SIZE=$ENV{WM_LABEL_SIZE})
-add_compile_options(-Wall -Wextra -Wold-style-cast -Wnon-virtual-dtor -Wno-unused-parameter -Wno-invalid-offsetof -Wno-attributes)
+add_compile_options(
+    -Wall -Wextra -Wold-style-cast -Wnon-virtual-dtor -Wno-unused-parameter -Wno-invalid-offsetof -Wno-attributes
+)
 add_compile_options(-O3)
 add_compile_definitions(NoRepository)
 if(CMAKE_COMPILER_IS_GNUCXX)
@@ -104,27 +106,27 @@ macro(test_OF_library)
     message(STATUS "Testing lib${OF_LIB_NAME}")
     add_executable(${OF_TEST_NAME} ${OF_TEST_DIR}/${OF_TEST_NAME}.C)
     add_test(${OF_TEST_NAME} 
-            COMMAND ${OF_TEST_NAME}
-            WORKING_DIRECTORY ${CMAKE_CURRENT_LIST_DIR}/${OF_TEST_DIR}/case
+             COMMAND ${OF_TEST_NAME}
+             WORKING_DIRECTORY ${CMAKE_CURRENT_LIST_DIR}/${OF_TEST_DIR}/case
     )
     # Include dirs
     foreach(loop_include_path ${OF_TEST_INCLUDE_PATHS})
         target_include_directories(${OF_TEST_NAME} PUBLIC ${loop_include_path})
-    endforeach()
+    endforeach(loop_include_path ${OF_TEST_INCLUDE_PATHS})
     # Linking
     target_link_libraries(${OF_TEST_NAME} LINK_PUBLIC ${Boost_LIBRARIES})
     foreach(loop_lib ${OF_TEST_LINK_LIBRARIES})
         target_link_libraries(${OF_TEST_NAME} LINK_PUBLIC ${loop_lib})
-    endforeach()
+    endforeach(loop_lib ${OF_TEST_LINK_LIBRARIES})
     target_link_libraries(${OF_TEST_NAME} LINK_PUBLIC OpenFOAM dl m)
     target_link_options(${OF_TEST_NAME} PUBLIC -fuse-ld=bfd)
     target_link_options(${OF_TEST_NAME} PUBLIC LINKER:--add-needed,--no-as-needed)
     add_custom_command(TARGET ${OF_TEST_NAME}
-                    POST_BUILD 
-                    COMMAND cp ${CMAKE_CURRENT_BINARY_DIR}/${OF_TEST_NAME} .
-                    COMMAND blockMesh && ./${OF_TEST_NAME} --log_level=message
-                    WORKING_DIRECTORY ${CMAKE_CURRENT_LIST_DIR}/${OF_TEST_DIR}/case                   
-                    COMMENT "Running ${OF_TEST_NAME}"
+                       POST_BUILD 
+                       COMMAND cp ${CMAKE_CURRENT_BINARY_DIR}/${OF_TEST_NAME} .
+                       COMMAND blockMesh && ./${OF_TEST_NAME} --log_level=message
+                       WORKING_DIRECTORY ${CMAKE_CURRENT_LIST_DIR}/${OF_TEST_DIR}/case                   
+                       COMMENT "Running ${OF_TEST_NAME}"
     )
 endmacro(test_OF_library)
 
[…]
