+++ .github/workflows/cmake.yml
@@ -32,31 +32,24 @@ jobs:
     steps:
     - uses: actions/checkout@v3
 
-    # - name: Cache OpenFOAM 9
-    #   id: cache-OF9
-    #   uses: actions/cache@v3
-    #   with:
-    #     path: |
-    #       /opt/openfoam9
-    #     key: ${{ runner.os }}-OF9
+    - name: Cache OpenFOAM 9
+      id: cache-OF9-CMake
+      uses: actions/cache@v3
+      with:
+        path: |
+          /opt/openfoam9
+          /usr/include/boost
+        key: ${{ runner.os }}-OF9
     
-    # - if: ${{ steps.cache-OF9.outputs.cache-hit == 'true' }}
-    #   name: OF9 cache hit
-    #   continue-on-error: true
-    #   run: |
-    #     echo OpenFOAM 9 already installed
+    - if: ${{ steps.cache-OF9-CMake.outputs.cache-hit == 'true' }}
+      name: OF9 cache hit
+      continue-on-error: true
+      run: |
+        echo OpenFOAM 9 already installed
+        sudo apt-get update
 
-    # - if: ${{ steps.cache-OF9.outputs.cache-hit != 'true' }}
-    #   name: Install OpenFOAM 9
-    #   shell: bash
-    #   run: |
-    #     sudo sh -c "wget -O - https://dl.openfoam.org/gpg.key > /etc/apt/trusted.gpg.d/openfoam.asc"
-    #     sudo add-apt-repository http://dl.openfoam.org/ubuntu
-    #     sudo apt-get update
-    #     sudo apt-get -y install openfoam9
-    #     sudo apt-get -y install cmake libboost-all-dev
-
-    - name: Install OpenFOAM 9
+    - if: ${{ steps.cache-OF9-CMake.outputs.cache-hit != 'true' }}
+      name: Install OpenFOAM 9
       shell: bash
       run: |
         sudo sh -c "wget -O - https://dl.openfoam.org/gpg.key > /etc/apt/trusted.gpg.d/openfoam.asc"
@@ -64,12 +57,24 @@ jobs:
         sudo apt-get update
         sudo apt-get -y install openfoam9
         sudo apt-get -y install cmake libboost-all-dev
+
+    # - name: Install OpenFOAM 9
+    #   shell: bash
+    #   run: |
+    #     sudo sh -c "wget -O - https://dl.openfoam.org/gpg.key > /etc/apt/trusted.gpg.d/openfoam.asc"
+    #     sudo add-apt-repository http://dl.openfoam.org/ubuntu
+    #     sudo apt-get update
+    #     sudo apt-get -y install openfoam9
+    #     sudo apt-get -y install cmake libboost-all-dev
     
     - name: cmake
       shell: bash
       continue-on-error: true
       run: |
         source /opt/openfoam9/etc/bashrc || true
+        mkdir -p $FOAM_RUN
+        mkdir -p $FOAM_USER_LIBBIN
+        mkdir -p $FOAM_USER_APPBIN
         cmake -S . -B build
         cd build
         make
