+++ docs/source/tutorials/Subroto2021.rst
@@ -54,7 +54,7 @@ Generate the blockMeshDict file using ``Subroto2021.system.cylinder.write_blockM
 
 The axisymmetryc mesh consists of a slice of a cylinder that matches the billet dimensions. The z_points tuple list the vertices z coordinates as defined in :numref:`lebon2020_bc`. The patch names are entered in the faces list in the ``Lebon2020.system.cylinder.write_boundary()`` function.
 
-Return to the case directory and generate the mesh using `blockMesh <https://doc.cfd.direct/openfoam/user-guide-v6/blockmesh>`_.
+Return to the case directory and generate the mesh using `blockMesh <https://doc.cfd.direct/openfoam/user-guide-v9/blockmesh>`_.
 
 .. code-block:: console
 
