+++ docs/source/tutorials/Lebon2020.rst
@@ -53,7 +53,7 @@ Generate the blockMeshDict file using the ``Lebon2020.system.cylinder.write_bloc
 
 The 3D mesh consists of a finer inner section that is definer by a square length of 32 mm but with each side with a slight curvature of radius 40 mm. The outer cylinder corresponds to the billet dimensions. The z_points tuple list the vertices z coordinates as defined in :numref:`lebon2020_bc`. The patch names are entered in the faces list in the ``Lebon2020.system.cylinder.write_boundary()`` function.
 
-Return to the case directory and generate the mesh using `blockMesh <https://doc.cfd.direct/openfoam/user-guide-v6/blockmesh>`_.
+Return to the case directory and generate the mesh using `blockMesh <https://doc.cfd.direct/openfoam/user-guide-v9/blockmesh>`_.
 
 .. code-block:: console
 
