+++ _sources/tutorials/Lebon2020.rst.txt
@@ -1,0 +1,397 @@
+===============================================================================
+Tutorial: Lebon *et al.* (2020) - 3D model of the DC casting of an A6060 billet
+===============================================================================
+
+.. contents:: Contents:
+  :backlinks: none
+
+This tutorial describes the steps to pre-process, run and post-process a 3D direct chill (DC) casting case that corresponds to the A6060 conventional DC casting experiment described in `Lebon et al. (2020) <https://doi.org/10.1016/j.apm.2019.08.032>`_ with directChillFoam.
+
+.. figure:: ../images/DC_Schematic.png
+  :width: 80%
+  :alt: DC casting schematic
+
+  The DC casting process: the melt in contact with the mould solidifies and the billet is pulled downwards by gravity as the ram is lowered. Water is sprayed at the sides of the billet providing the direct chill.
+
+The execution time of this tutorial case is around 5 minutes when run in parallel with 80 cores on two nodes with 2.50 GHz processors.
+
+Pre-processing
+==============
+
+The schematic for the case setup is as follows:
+
+.. _Lebon2020_BC:
+
+.. Figure:: ../images/Lebon2020_BC.png
+  :width: 80%
+  :alt: Schematic of the Lebon *et al.* (2020) case
+
+  Boundary patch labels and dimensions of the casting mould and billet. The image is not drawn to scale.
+
+The first step is to generate the mesh for the case and setup the boundary conditions.
+
+To run the pre-processing steps, change to the case directory.
+
+.. code-block:: console
+
+  $ cd directChillFoam/tutorials/heatTransfer/directChillFoam/Lebon2020
+
+Mesh generation
+---------------
+
+The dictionary file that describes the mesh is located in the system directory.
+
+.. code-block:: console
+
+  $ cd system
+
+Generate the blockMeshDict file using the ``Lebon2020.system.cylinder.write_blockMeshDict()`` function from the `system/cylinder.py script <modules.html#id6>`_.
+
+.. code-block:: console
+
+  $ python3 cylinder.py
+
+The 3D mesh consists of a finer inner section that is definer by a square length of 32 mm but with each side with a slight curvature of radius 40 mm. The outer cylinder corresponds to the billet dimensions. The z_points tuple list the vertices z coordinates as defined in :numref:`lebon2020_bc`. The patch names are entered in the faces list in the ``Lebon2020.system.cylinder.write_boundary()`` function.
+
+Return to the case directory and generate the mesh using `blockMesh <https://doc.cfd.direct/openfoam/user-guide-v9/blockmesh>`_.
+
+.. code-block:: console
+
+  $ cd ..
+  $ blockMesh
+
+Boundary and initial conditions
+-------------------------------
+
+To speed up the simulation, initialize the fields with rough estimates of temperatures (and corresponding liquid fractions) in the case directory:
+
+.. code-block:: console
+
+  $ setFields
+
+.. warning::
+  The liquid fraction values must match the temperatures for the alloy according to the constant/tStar Foam table. Otherwise, the case can diverge.
+
+Solute
+^^^^^^
+
+Solute fields are dimensionless: 
+
+.. code-block:: C
+
+  dimensions      [0 0 0 0 0 0 0];
+
+The solute fields are initialized with constant values. Use 0.6 for silicon and 0.48 for magnesium.
+
+.. code-block:: C
+
+  internalField   uniform 0.45; // C.Si
+
+.. code-block:: C
+
+  internalField   uniform 0.475; // C.Mg
+
+The inlet boundary condition is a fixed value (stored in internalField):
+
+.. code-block:: C 
+  
+  free-surface
[…]
