### 2023-02-16 6a25508
JOSS editor review: Typo and references
