+++ paper/paper.bib
@@ -20,7 +20,7 @@
 
 @article{Bennon1987,
   author  = {Bennon, W. D. and Incropera, F. P.},
-  doi     = {https://doi.org/10.1016/0017-9310(87)90094-9},
+  doi     = {10.1016/0017-9310(87)90094-9},
   issn    = {0017-9310},
   journal = {International Journal of Heat and Mass Transfer},
   number  = {10},
@@ -223,7 +223,7 @@
 
 @article{Vreeman2000,
   author  = {Vreeman, Christopher J. and Krane, Matthew John M. and Incropera, Frank P.},
-  doi     = {https://doi.org/10.1016/S0017-9310(99)00174-X},
+  doi     = {10.1016/S0017-9310(99)00174-X},
   issn    = {0017-9310},
   journal = {International Journal of Heat and Mass Transfer},
   number  = {5},
