### 2023-02-16 #4871 @blebon
@meg-simula Thank you for editing this submission. I have fixed the typo in the introduction and fixed the references that the editorial bot complained about above.

Regarding your final comments:

- The new tag is OF9.0.3-JOSS: https://github.com/blebon/directChillFoam/releases/tag/OF9.0.3-JOSS
- This tag is archived in Zenodo at https://zenodo.org/record/7645448
- I have updated the metadata in Zenodo. I am the sole author of this software, so the authors list is correct.
- The doi of the archive is [10.5281/zenodo.7645448](https://dx.doi.org/10.5281/zenodo.7645448)
