### 2023-02-16 #4871 @kyleniemeyer
Hi @blebon, I made a few minor edits to the paper, can you merge this PR? https://github.com/blebon/directChillFoam/pull/2

After that we are good to publish.
