### 2023-02-19 #4871 @arfon
Thanks again everyone, and especially to @meg-simula who is stepping down from the JOSS editorial team. Thank you for all of your work to support JOSS @meg-simula 💖
